+++ IPython/kernel/inprocess/blockingkernelmanager.py
