+++ IPython/kernel/inprocess/socket.py
