+++ IPython/kernel/tests/test_kernelmanager.py
