+++ IPython/kernel/zmq/displayhook.py
@@ -2,7 +2,7 @@ import __builtin__
 import sys
 
 from IPython.core.displayhook import DisplayHook
-from IPython.inprocess.socket import SocketABC
+from IPython.kernel.inprocess.socket import SocketABC
 from IPython.utils.jsonutil import encode_images
 from IPython.utils.traitlets import Instance, Dict
 from session import extract_header, Session
